feat(angular): add addModuleSpec flag to library schematic + refactor
